% Auto-generated: do not edit by hand
\name{cloveriiif}

\alias{cloveriiif}

\title{Cloveriiif component}

\description{
The Cloveriiif component takes one parameter 'iiifLink'. If an iiifLink is not passed in, then it will be set to the default image "Crossing the Pend d'Oreille - Kalispel".
}

\usage{
cloveriiif(iiifLink=NULL)
}

\arguments{
\item{iiifLink}{Character. The link to the iiif image.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

